Мелкие исправления

+++ private/classes/FileManager/File.php
@@ -91,9 +91,11 @@ class File {
      * @access public
      */
     public function delete() {
-        $result = unlink($this->_path . $this->_name);
-        if ($result === false) {
-            throw new Exception('Can not delete file ' . $this->_name);
+        if (file_exists($this->_path . $this->_name)) {
+            $result = unlink($this->_path . $this->_name);
+            if ($result === false) {
+                throw new Exception('Can not delete file ' . $this->_name);
+            }
         }
     }
